requirements.txt
kafka-python == 2.2.15
elasticsearch == 8.12.0
websockets == 13.1